claude/ticket-management/backend/auth/requirements.txt
# Authentication Service Requirements
# Reference: /backend/auth/README.md

# FastAPI and Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pyotp==2.9.0  # For MFA/TOTP
qrcode==7.4.2  # For MFA QR codes

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Environment & Configuration
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2

# Testing (Optional)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0  # For async HTTP client in tests

# Code Quality (Optional)
black==24.1.1
flake8==7.0.0
mypy==1.8.0
